Improve localisation for radio options
Extend the localisation schema with the `value` of the radio option (for example "yes" or "no") to be able to localise these automatically.

Also avoid double lookups in method `localised_text`. No need to first check if a key exists, if we pass to the translate method a `nil` default. It will behave the same but make only one lookup.

This is a backwards compatible change so the interface to the `#govuk_collection_radio_buttons` method does not change, which means to be able to take the label from the locales, instead of providing to the method a `text_method`, just pass `nil`.

Example:

```ruby
f.govuk_collection_radio_buttons :parent, @values, :value, nil, inline: true
```

This makes for a bit ugly interface but could be improved later on. I didn't want to break compatibility.