Improve localisation further and allow html locales
I've done a small refactor so the collection check boxes hint elements are the same as with other form elements, by using the same class and method signature.

I couldn't find a reason why this was not the case, probably it was a leftover from a previous refactor.

Also, now localisation can also contain html elements, and in order to follow the usual Rails conventions, these locale keys will contain the suffix `_html` which will make the content html safe automatically when using `t()` method in the view, although it does not happen automatically outside the view that's why I use `#html_safe` if a locale was found.

Still, the non-html key will be tried first, and returned if one is found, otherwise the html version will be tried, and if none was found, nil will be returned as before.